requirements.txt
aiohttp==3.7.4.post0
aioresponses==0.7.2
async-timeout==3.0.1
attrs==21.2.0
beautifulsoup4==4.9.3
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
httpretty==1.1.3
idna==2.10
idna-ssl==1.1.0
ipinfo==4.2.0
lxml==4.6.3
multidict==5.1.0
Pillow==8.2.0
python-nmap==0.6.4
python-Wappalyzer==0.3.1
requests==2.25.1
soupsieve==2.2.1
typing-extensions==3.10.0.0
urllib3==1.26.5
yarl==1.6.3
